zenquotes_emailer/requirements.txt
requests
python-dotenv
schedule
yagmail
psycopg2-binary
